// Dashboard/src/types.ts
export interface EVData {
  VIN: string;
  County: string;
  City: string;
  State: string;
  PostalCode: string;
  ModelYear: number;
  Make: string;
  Model: string;
  ElectricVehicleType: string;
  CleanAlternativeFuelVehicle: string;
  ElectricRange: number;
  BaseMSRP: number;
  LegislativeDistrict: number;
  DOLVehicleID: number;
  VehicleLocation: string;
  ElectricUtility: string;
  CensusTract: number;
}